third_party/WebKit/LayoutTests/typedcssom/the-stylepropertymap/computed-expected.txt

This is a testharness.js-based test.
FAIL Computed StylePropertyMap contains every CSS property assert_equals: expected 298 but got 296
PASS Computed StylePropertyMap reflects CSS property declarations in style rules
FAIL Computed StylePropertyMap reflects custom property declarations in style rules assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
PASS Computed StylePropertyMap reflects CSS property declarations in inline styles
FAIL Computed StylePropertyMap reflects custom property declarations in inline rules assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
PASS Computed StylePropertyMap contains computed values and not resolved values
PASS Computed StylePropertyMap is live
Harness: the test ran to completion.
